+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 1.2.0 - Spit Shine
+
+- Add `Logger.error(Element, AnnotationMirror, Object)`
+- Add `Logger.warn(Element, AnnotationMirror, Object)`
+- Add `Logger.note(Element, AnnotationMirror, Object)`
+- Add `TypeMirrors.annotation(Element, DeclaredType)`
+- Add `TypeMirrors.annotations(Element, DeclaredType)`
+- Change `Logger.error(Element, String)` to `Logger.error(Element, Object)`
+- Change `Logger.warn(Element, String)` to `Logger.warn(Element, Object)`
+- Change `Logger.note(Element, String)` to `Logger.note(Element, Object)`
+- Remove `Logger.error(Element, Object, String, String)`
+- Remove `Logger.error(Element, Object, String)`
+- Remove `Logger.warn(Element, Object, String, String)`
+- Remove `Logger.warn(Element, Object, String)`
+- Remove `Logger.note(Element, Object, String, String)`
+- Remove `Logger.note(Element, Object, String)`
+
 ## 1.1.1 - Polish & Shine (20/06/2021)
 
 This release focuses on fixing a few pesky issues.